# anyflix-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media Streaming Backend Service",
...
    "contact": {
      "name": "Media Streaming Backend API",
      "url": "https://github.com/your-repo/media-backend"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/sources/": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "📋 List Available Sources",
        "description": "Get all available media sources.",
        "operationId": "get_sources_sources__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sources/{source}/preferences": {
      "get": {
        "tags": [
          "media-api"
        ],
        "summary": "📋 Get Source Configuration",
        "description": "Get configuration preferences for a specific source.",
        "operationId": "get_source_preferences_sources__source__preferences_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sources/trailer": {
      "post": {
        "tags": [
          "media-api"
        ],
        "summary": "🎬 Extract Streamable Trailer URL",
        "description": "Extract streamable URL from AniList or TMDB trailer data.\n\nTakes trailer information from AniList or TMDB responses, builds the full YouTube URL,\nand uses ytdlp_extractor to get the actual streamable URL.\n\nArgs:\n    request: TrailerRequest containing either anilist_trailer or tmdb_trailer data\n\nReturns:\n    TrailerResponse with streamable URL and metadata",
        "operationId": "extract_trailer_url_sources_trailer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferencesResponse": {
        "properties": {
          "preferences": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preferences",
            "description": "Source configuration preferences"
          }
        },
        "type": "object",
        "required": [
          "preferences"
        ],
        "title": "PreferencesResponse",
        "description": "Response for source preferences/configuration."
      },
...
      "SourcesResponse": {
        "properties": {
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "List of available source names"
          }
        },
        "type": "object",
        "required": [
          "sources"
        ],
        "title": "SourcesResponse",
        "description": "Response for available sources list."
      },
...
      "TrailerRequest": {
        "properties": {
          "anilist_trailer": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anilist Trailer"
          },
          "tmdb_trailer": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Trailer"
          }
        },
        "type": "object",
        "title": "TrailerRequest",
        "description": "Request model for trailer extraction."
      },
      "TrailerResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "original_url": {
            "type": "string",
            "title": "Original Url"
          },
          "streamable_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Streamable Url"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "original_url"
        ],
        "title": "TrailerResponse",
        "description": "Response for trailer extraction."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}